--- a/dbms/src/Common/UnifiedLogFormatter.h
+++ b/dbms/src/Common/UnifiedLogFormatter.h
@@ -14,8 +14,9 @@
 
 #pragma once
 
+#include <Common/FmtUtils.h>
+#include <Poco/Formatter.h>
 #include <Poco/Message.h>
-#include <Poco/PatternFormatter.h>
 
 #include <string>
 
@@ -24,23 +25,41 @@ namespace DB
 class WriteBuffer;
 
 /// https://github.com/tikv/rfcs/blob/ed764d7d014c420ee0cbcde99597020c4f75346d/text/0018-unified-log-format.md
-class UnifiedLogPatternFormatter : public Poco::PatternFormatter
+class UnifiedLogFormatter : public Poco::Formatter
 {
 public:
-    UnifiedLogPatternFormatter() = default;
+    UnifiedLogFormatter() = default;
 
     void format(const Poco::Message & msg, std::string & text) override;
 
 private:
-    static std::string getPriorityString(const Poco::Message::Priority & priority);
+    enum class JsonEncodeKind
+    {
+        /**
+         * No need to encode, just copy the text
+         */
+        DirectCopy,
 
-    static std::string getTimestamp();
+        /**
+         * Add double quotes around the text is sufficient
+         */
+        AddQuoteAndCopy,
 
-    static bool needJsonEncode(const std::string & src);
+        /**
+         * Need full JSON string encode
+         */
+        Encode,
+    };
 
-    static void writeJSONString(DB::WriteBuffer & buf, const std::string & str);
+    static void writePriority(FmtBuffer & buf, const Poco::Message::Priority & priority);
 
-    static void writeEscapedString(DB::WriteBuffer & wb, const std::string & str);
+    static void writeTimestamp(FmtBuffer & buf);
+
+    static JsonEncodeKind needJsonEncode(const std::string & src);
+
+    static void writeJSONString(FmtBuffer & buf, const std::string & str);
+
+    static void writeEscapedString(FmtBuffer & buf, const std::string & str);
 };
 
 } // namespace DB