--- a/dbms/src/DataStreams/HashJoinBuildBlockInputStream.h
+++ b/dbms/src/DataStreams/HashJoinBuildBlockInputStream.h
@@ -31,7 +31,7 @@ public:
         size_t concurrency_build_index_,
         const String & req_id)
         : concurrency_build_index(concurrency_build_index_)
-        , log(Logger::get(NAME, req_id))
+        , log(Logger::get(req_id))
     {
         children.push_back(input);
         join = join_;