--- a/dbms/src/DataStreams/MockExchangeSenderInputStream.cpp
+++ b/dbms/src/DataStreams/MockExchangeSenderInputStream.cpp
@@ -18,7 +18,7 @@ namespace DB
 MockExchangeSenderInputStream::MockExchangeSenderInputStream(
     const BlockInputStreamPtr & input,
     const String & req_id)
-    : log(Logger::get(NAME, req_id))
+    : log(Logger::get(req_id))
 {
     children.push_back(input);
 }