@fuzhe1989 2022-10-08
I'm curious about whether we could treat every log as a static message plus multiple key-value pairs.