@breezewish 2022-10-08
> I'm curious about whether we could treat every log as a static message plus multiple key-value pairs.

Definitely we could! It is just not efficient in Poco due to heavy allocations :) For example, adding a single field will +300ns in my M1 (which means roughly 50% slow down)