@fuzhe1989 2022-10-08
@breezewish I don't mean it should take effect right now. Maybe first provide a key-value style api is very helpful.

Besides, [spdlog](https://github.com/gabime/spdlog) could be a competitive alternative of Poco log.